refactor(Lotto): lotto와 lottoPrize 패키지로 코드 구조 정리
